requirements.txt
numpy>=1.20.0
pyserial>=3.5
matplotlib>=3.6
seaborn>=0.12
PyYAML>=6.0
joblib>=1.3
scikit-learn>=1.2
scipy>=1.10
pandas>=2.0
fastapi>=0.100
uvicorn[standard]>=0.20
websockets>=11.0
skl2onnx>=1.15
onnxruntime>=1.15
